Added support for links with delay.
The delay parameter is taken from the contact plan (contact topology)
and it is applyed by the Com layer. It is important to configure the
ranges in the contact file (in light seconds unit). CGR-based routing
routines were updated to take the One-Way Light Time (owlt) estimation
into account in route calculations as well.